drm/fourcc: fix cpp interpretation when using char_per_block
Commit 9a29ec15a72b ("drm/fourcc: Add 32/64x4 Tile formats support")
changed drm_format_info::char_per_block to u32, while keeping cpp as
u8 within the same union. This introduced an inconsistency for formats
that still rely on cpp to store bytes per pixel in `__drm_format_info`.

As a result, `drm_format_info_min_pitch()` could return incorrect
values by reading from the u32 char_per_block field instead of the
intended cpp value.

Fix this by updating cpp to u32 as well, ensuring both fields are
consistently aligned within the union and interpreted correctly.

Fixes: 9a29ec15a72b ("drm/fourcc: Add 32/64x4 Tile formats support")
Message-ID: <20260406100903.3502861-1-neel.gandhi@amd.com>
State: pending

+++ include/drm/drm_fourcc.h
@@ -83,7 +83,7 @@ struct drm_format_info {
 		 * triplet @char_per_block, @block_w, @block_h for better
 		 * describing the pixel format.
 		 */
-		u8 cpp[DRM_FORMAT_MAX_PLANES];
+		u32 cpp[DRM_FORMAT_MAX_PLANES];
 
 		/**
 		 * @char_per_block:
